refactor: Moved more methods into scatter_plot_data_utils.ts
